requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
sentence-transformers==3.2.1
numpy==1.26.4
python-dotenv==1.0.1
pytest==8.3.3
